Update Public Secret
I had to modify this code because this answer:

class Person 
  attr_accessor  :secret

  def initialize; @secret = nil; end
end

It can be simplified. Initializing an instance variable @secret to nil is unnecessary because I am setting an instance variable secret to a value immediately after instantiating the Person object. Because of this, I do not need to create a constructor method, because the setter method from attr_accessor will instantiate an instance variable @secret to the specified value automatically.